backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
azure-cosmos==4.9.0
azure-identity==1.19.0
azure-search-documents==11.6.0
azure-ai-inference==1.0.0b6
openai==1.55.3
python-multipart==0.0.18
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
httpx==0.28.1
azure-storage-blob==12.23.1
azure-mgmt-resource==23.2.0
azure-mgmt-costmanagement==4.0.1
azure-mgmt-monitor==6.0.2
tiktoken==0.5.2
pdf2image==1.17.0
Pillow==10.4.0
beautifulsoup4==4.12.3
requests==2.32.3
PyPDF2==3.0.1
